Big file restructuring on server-side code
