--- server/task/delegateTasksTest.js
@@ -1,22 +1,0 @@
-var taskModel = require('./taskModel');
-var userModel = require('../user/userModel');
-
-
-
-
-
-var dwellingId = 1;
-
-//getAll Instances
-taskModel.getAllInstances(dwellingId, function (err, taskInstances){
-  userModel.getByDwellingId(dwellingId, function (err, users){
-    taskModel.delegateInstances(users, taskInstances, function (err, results){
-
-    })
-  })
-})
-
-
-
-
-
